[WIP] Add a basic transaction view (#159)
* webapp: add transaction classifier to tx summary page

The body is somewhat verbose, but it's made that way to be extensible to more
classifications in the future.

* fonts: add iosevka, use it consistently for cryptographic data

This uses Iosevka Term rather than Iosevka, so that all Unicode characters have
a consistent width, even the "weird" ones. I picked one of the stylistic
variants to reduce the size. At some later point we could build our own font
files if we cared about the weight.

* webapp: add a transaction view

This code probably has a bunch of structuring issues and could be refactored to
make it better, but it sets up being able to completely render every aspect of
a `TransactionView`.  We can also probably reuse some of the components for
rendering `TransactionPlan`s in the confirmation dialog.

* add todo on viewbox

* right-align destinations in send/output

* add sketch of toggleable view

* Rebase & Refactor

* Refactor address view component

---------

Co-authored-by: Gabe Rodriguez <grod220@gmail.com>