wasm: simplify transaction planner
Closes #609

This commit gets most of the way to dramatically simplifying the transaction
planner.  It completely eliminates the fork of the Rust crate's `Planner` API,
and focuses on what the web extension actually needs, namely, serving
`TransactionPlanRequest`s.

The only thing remaining is cleaning up the test code, which was written
against the previous API and much of which doesn't make sense any more.

Along the way, it also fixes a bug in the existing web extension code, which
would result in making unnecessary spends sometimes (always spending all
available notes, even if unnecessary).

One observation growing out of this (cc @zbuc) is that we should change the way
gas costs are computed on-chain, removing the transaction-wide byte costs and
making the gas cost of the transaction just be the sum of the gas costs of its
constituent actions. (In other words, we'd attribute the block space costs of
each action to that action, and make the surrounding transaction contents free,
rather than what we currently do, which is have the block space costs
attributed to the transaction as a whole and make it free for each action). The
reason is that this makes it very difficult to do "incremental" reasoning about
the gas costs of the transaction as actions are added or removed.

We could also later backport much of this simplification to the Rust planner.