build: make zip file exports deterministic (#2098)
As long as the zip files are exported from the same git commit, the
resulting archives should be byte-for-byte identical. Having fully
reproducible builds will require comprehensive dependency pinning,
e.g. via nix flake or similar, to ensure that the wasm artifacts are
built identically, but for now, simply clobbering the timestamps of the
exported asset files provides decent reproducibility.