ci: update deployment workflows for dex-explorer (#2147)
## Description of Changes

Follow-up to https://github.com/penumbra-zone/web/pull/2134. Ports the
CI workflows to the repo root, so they'll run on the right events.

The substantive changes here are to the `Containerfile`. Given the
migration of Veil to a monorepo setup, we now have a much larger build
context to handle: we must build all relevant dependencies prior to
building Veil, including using Rust tooling. That's handled now. The
resulting image is larger: ~1.2GB compared to ~300MB for the old
dex-explorer repo, but I'll punt on further refinements until we have CD
flowing again.

## Related Issue
https://github.com/penumbra-zone/web/pull/2134

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.