feat: minifront_version_in_footer_is_a_broken_link (#2100)
## Description of Changes

I did some research and found that the "GIT_ORIGIN_URL" variable gets
its value from "execSync('git remote get-url origin')". This works fine
when the repo is cloned using HTTPS, but if it's cloned via SSH, it
returns something like
[git@github.com](mailto:git@github.com):username/reponame. When this SSH
URL is used in an tag, the browser treats it as a relative URL, which is
why it gets prefixed with "http://localhost/:.."

To fix this, I detect the fetched URL, and if it is in SSH format, I
convert it to the corresponding HTTPS format.

Please provide a clear and concise description of the changes made in
this PR.
For UI changes, include images or videos to show the visual impact.

## Related Issue
Closes #2057 
Link to the issue this PR addresses.

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.