ci: use nix devshell for workflows
We want a single declaration of tooling dependencies required to work on
the apps in the web monorepo. Extending the nix flake so that it's used
in CI, as well, ensuring that CI failures are easily replicable.

There's a single omission, for the sake of velocity: I haven't ported
the playwright tests to use the nix env. That'll take another pass, but
punting on it for now, so that we have green CI again on the web
monorepo.