Social card dialog (1/2) (#2256)
## Description of Changes

Implements https://github.com/penumbra-zone/web/issues/2241 (with
trigger pending in a follow up PR)

Context: When we click on Share to X, we can't directly upload an image
in the post. So the workaround is the create a unique link with a
twitter:img tag, so that the link will contain the social image. To make
this work, we need to be able to generate the image on the server-side.
That's what this PR does in addition to the dialog.

To summarize, it contains:
- The social card dialog
- A page that contains the twitter metatags so that it can be shared on
x.com (which redirects to the tournament page)
- Serverside rendering of the canvas using an api route

<img width="1197" alt="image"
src="https://github.com/user-attachments/assets/f35ebdb8-57fd-4ba2-9f57-74239354b8cf"
/>

<img width="639" alt="image"
src="https://github.com/user-attachments/assets/acbf6774-e144-4d67-be31-6d94209f0d77"
/>

## Related Issue

Link to the issue this PR addresses.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>
Co-authored-by: Lucas Meier <lucas@cronokirby.com>