Improve registry / asset fetching (#2336)
## Description of Changes

This relies solely on our registry for asset fetching. This closes #2295
This also greatly simplifies the client side code, and improves
performance, by fetching the registry once on the server side (atm, for
any page), and including it in the bundle, avoiding the need for an
additional round trip on the client, and removing code which needlessly
checks whether or not the registry is loaded, by guaranteeing that it
always is loaded.

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.