veil: bug: distinguish no lp rewards and loading (#2349)
The bug here is that we were treating the case where we knew the user
had no positions and the case where we were waiting to learn if the user
had any positions identically, causing an infinite load.

## Description of Changes

This handles things correctly now, returning an empty rewards response
when you have no positions.

Closes #2348.
## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.