fix(perspective): #2280: tx summary swap rendering (#2369)
Closes #2280

Applies correct sides to assets of swap and swap claim action views to
transaction summary component, as the issue suggests.

Doesn't do anything with issue's claim of "the swap only shows input
amount and no output amount" – that's an expected behavior of the swap
view – output amount is shown in the swap claim summary.