Wire up specific delegator page (#2368)
## Description of Changes

Closes #2343.

The logic here ended up needing to be substantially different than the
existing code path, which was tightly coupled to fetching rewards for a
specific delegator.

I also took the freedom to do some light refactoring, and to try out
"use server" as a way to avoid the tedium of writing out an API route. I
think it works pretty nicely!

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.