ci: playwright symlink madness
It's absolutely cursed what I'm doing here, but the goal was:

  * get playwright tests running under linux, so i can run them locally
  * use the nix env to provide tooling
  * ergo: run the playwright tests within the devshell

Surprisingly difficult! Playwright demands binaries in a very specific
location, and the binaries it downloads are dynamically linked, to paths
that will be filtered out of the nix devshell's LD_LIBRARY_PATH, and
therefore the browsers won't run. We can install suitable browsers
within nix and cache them, but instructing Playwright to use them is
actually quite difficult. So we cheat, and create symlinks at the
paths Playwright watches, pointing to the nix-downloaded browsers.

Cursed? You bet. Does it work? It does! Although it'll break whenever
the `playwright` node deps are bumped.