chore: bump wasm deps to v2.0.0-alpha.11 (#2408)
## Description of Changes
Bumps the wasm dependencies to the latest LQT-related tag, which
includes a rebase on `v1.5.2` changes. No anticipated breakage for web
code; merely updating to keep things well in sync, in anticipation of
final release.

Refs:

  * https://github.com/penumbra-zone/penumbra/issues/5203
  * https://github.com/penumbra-zone/penumbra/issues/5010

## Related Issue

See above.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.

I have done no such thing. Also, that reminds me: we're due for a
minifront update in the protocol repo. Will do that again ahead of
release.