ci: add cache for playwright install (#2384)
## Description of Changes

Adds a cache step specifically for the playwright browser installation
step, so it doesn't need to download the browsers on every CI run. This
is intentionally a minimal diff, favoring improvement over perfection.
Future work should refactor the turbo workflows with an eye toward
effective caching.

Closes #2366.

## Related Issue

Link to the issue this PR addresses.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.