feat: minifront v2 app (#2416)
Separated Minifront v2 from the classic Minifront because they are
mostly incompatible and the old version limits the freedom of movement.

The idea is the same as a separation of `ui` package from
`ui-deprecated`. When v2 is complete, we'll just delete the old version
together with `ui-deprecated`. This new version allows us to use new
tools and practices like MobX.

Follow up work: make it actually work. This PR is done in a way to keep
it simple and mostly move the files.