dev: support nix env for playwright tests
It's absolutely cursed what I'm doing here, but the goal was:

  * get playwright tests running under linux, so i can run them locally
  * use the nix env to provide tooling
  * ergo: run the playwright tests within the devshell

Surprisingly difficult! Playwright demands binaries in a very specific
location, and the binaries it downloads are dynamically linked, to paths
that will be filtered out of the nix devshell's LD_LIBRARY_PATH, and
therefore the browsers won't run. Even if we override the devshell's
library path, we'd still be referencing incompatible shared objects.

We can instead install suitable browsers within nix and cache them,
although instructing Playwright to use them is actually quite difficult.
So we cheat, and create symlinks at the paths Playwright watches,
pointing to the nix-downloaded browsers.

Cursed? You bet. Does it work? It does! Although it'll break whenever
the `playwright` node deps are bumped. Made top-level var in the wrapper
script so that updating the playwright browser version string is
straightforward.