dev: support nix env for playwright tests (#2419)
## Description of Changes

It's somewhat cursed what I'm doing here, but the goal was:

* get playwright tests running under linux, so i can run them locally
* use the nix env to provide tooling
* ergo: run the playwright tests within the devshell

Surprisingly difficult! Playwright demands binaries in a very specific
location, and the binaries it downloads are dynamically linked, to paths
that will be filtered out of the nix devshell's LD_LIBRARY_PATH, and
therefore the browsers won't run. Even if we override the devshell's
library path, we'd still be referencing incompatible shared objects.

We can instead install suitable browsers within nix and cache them,
although instructing Playwright to use them is actually quite difficult.
So we cheat, and create symlinks at the paths Playwright watches,
pointing to the nix-downloaded browsers.

Cursed? You bet. Does it work? It does! Although it'll break whenever
the `playwright` node deps are bumped. Made top-level var in the wrapper
script so that updating the playwright browser version string is
straightforward.

## Related Issue

Closes #2411. Refs #2410. 

## Testing and reviewing

In this changeset I've _avoided_ updating the behavior of CI, as
assurance that the changes I've made don't break the existing playwright
tests. In future work (#2410) I'll revisit CI and pull this logic in.

If you're reviewing this PR on a mac, verify that nothing in it affects
you. If you're reviewing this PR and you use a non-[Playwright-supported
Linux distro](https://playwright.dev/docs/intro#system-requirements),
and you feel comfortable using a nix devshell, go ahead and blast `just
test` and see how it treats you.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.

I've not done so, but was careful not to alter actual app code when
preparing the diff.