services: make tests use type-correct idb mock (#2373)
Replaces the arbitrarily-declared `IndexedDbMock` in the services
package with a preconstructed mock object that enforces the actual
interface type.

This is important because now the mock no longer has to be manually
maintained in parallel to the storage interface - it updates
automatically when the type updates, requiring the tests to also update.