ci: leverage turborepo caching (#2421)
## Description of Changes

Reduces CI runtime on the primary "Turbo CI" workflow, by:

1. careful management of nix devshell for tooling
2. careful management of `.turbo/` directory for turbo cache
3. removing unnecessary workflow serialization

Rerunning the "Turbo CI" workflow on this changeset shows a 2-3x
reduction in wall time:

* [Run
1](https://github.com/penumbra-zone/web/actions/runs/15219766711), 6m5s
* [Run
2](https://github.com/penumbra-zone/web/actions/runs/15219872779), 6m23s
* [Run
3](https://github.com/penumbra-zone/web/actions/runs/15220037312), 6m11s

Whereas previous runs were clocking in at ~15-20m. See more detailed
timing info in #2410.

## Related Issue

Closes #2410. 

## Testing and review

First, review and merge https://github.com/penumbra-zone/web/pull/2419,
since this PR includes that changeset, as well. Once that PR is merged,
the diff here will appear much smaller.

It's worth pointing out there there's still a lot of duplicative work
done in the parallel CI jobs, but I'm fine with that: CI minutes are
cheap, and preserving the UX of distinct lint and test jobs strikes me
as beneficial during a period of heavy churn on this repo.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.

No, I've not done that, but the bulk of this diff applies to CI, so I'm
not worried about breakage.