build: support build-time BASE_URL (#2447)
## Description of Changes

Here we introduce a container build-arg that permits overriding the
`NEXT_PUBLIC_BASE_URL` env var for the nextjs prod build. This is not a
comprehensive solution: there's a slew of `process.env` calls scattered
around the Veil codebase. Cleaning up the interface for obtaining
runtime config in client-side can wait, however; right now, I'm shoving
things around to keep the deploys flowing in their current forms, so
that hotfixes to Veil land immediately.

## Related Issue

Refs #2436.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.