LP leaderboard: fix asset selector (#2474)
## Description of Changes

1. Originally, the asset selector in LP leaderboard (round page)
displayed the user's balance next to the complete list of assets. On the
page where it was implemented, I initially tried replacing it with the
VotingAssetSelector to match a similar structure, but that didn’t quite
work out. Instead, I decided to combine elements from both approaches
and ultimately kept the `AssetSelector` from the UI package, with some
modifications:

```ts 
   <AssetSelector
      assets={metadataAssets}
      balances={undefined}
      value={selectedAsset}
      onChange={setSelectedAsset}
   />
```

Set the balances entry to undefined (since we're only interested in the
assets), and filter for the gauge assets by calling `useEpochResults`
API, then pass those directly into the `AssetSelector`.

2. Introduces smoother animations and transitions, addressing
https://github.com/penumbra-zone/web/pull/2401#pullrequestreview-2858362541
when refreshing the page (which collapses the table), switching between
assets in the asset selector, and toggling between "All LPs" and "My
LPs" tabs.

3. Updates the message from "You have no liquidity positions in this
epoch" to "Your LPs have not received any rewards during this epoch" to
avoid misleading language.

_However_, this doesn't address the feedback: "there should also not be
an 'all assets' view for the LP leaderboard..." — what should we be
showing instead? A default asset in the asset selector on page load? Is
it fine to keep this unidirectionality? cc @garwalsh @@smmhrdmn

---------------


https://github.com/user-attachments/assets/9d5d1656-8b40-4877-a7d9-f779ec222bf1

## Related Issue

references https://github.com/penumbra-zone/web/issues/2471

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.