fix(veil): #2491: filter positions by status correctly (#2492)
Closes #2491 

It's not ideal but it does the job – moved filtering by status to
request-level instead of client filtering (it was happening too late
after pagination).

Technically the pagination is not correct as it can load 40 items
instead of 20 (BASE_LIMIT), and so what?:) We have terrible data
modeling around positions and position ids – to fix it properly we need
to rethink the protos.