positions: display proper position states in the tables (#2495)
## Description of Changes

follow-up to https://github.com/penumbra-zone/web/pull/2492

currently, `PositionsTable` performs state filtering internally. Instead
of filtering based on the `showInactive` boolean, we move the state
logic up to the calling component, and update `PositionsTable` to accept
an explicit array of `PositionState_PositionStateEnum` proto type
values. This fixes a regression that caused open positions to appear in
the closed positions table on the portfolio page.

---------


https://github.com/user-attachments/assets/d9620d53-f939-4736-8469-5406b93fc40d


## Related Issue

N/A

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.