feat: #2004: tailwind eslint (#2512)
This PR moves from using
[eslint-plugin-tailwindcss](https://github.com/francoismassart/eslint-plugin-tailwindcss)
to
[eslint-plugin-better-tailwindcss](https://github.com/schoero/eslint-plugin-better-tailwindcss)
because the older plugin [doesn't support Tailwind
v4](https://github.com/francoismassart/eslint-plugin-tailwindcss/issues/325).

This PR is an addition to #2511. It is split to not overflow the
mentioned PR with too many auto-generated changes.

Here, eslint-plugin-better-tailwindcss is applied to UI package, Veil
and Minifront v2 apps, because these are the only constantly-updating
packages that require Tailwind v4.

Main change is here:
https://github.com/penumbra-zone/web/pull/2512/files#diff-a16f7ad5bcb64913244872ef7f7eb8932dc063d76feb2336070098e6963bfab5R1.
The rest files simply import this function and perform `pnpm lint` in
the mentioned packages.