try: fix not being able to vote (#2524)
## Description of Changes

The UI logic prevents you from voting if you can't load the metadata of
the first voting note's delegation asset. This is only because you need
to render a "generic" delegated UM token. This PR instead bases the
generic metadata off of the UM metadata, so the icon is the same as UM,
but with a different symbol.

## Related Issue

Closes #2522.

Link to the issue this PR addresses.

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.