Fix Price Ranges (#2528)
## Description of Changes

Closes: https://github.com/penumbra-zone/web/issues/2529

There's a problem where the price range initializes before the market
price that causes the handles to go out of bounds:

![image](https://github.com/user-attachments/assets/30827d18-5610-423f-892c-4cfda75f0446)

This PR fixes it by only rendering the price range once the market price
has been fetched.

## Related Issue

Link to the issue this PR addresses.

## Checklist Before Requesting Review

- [ ] I have ensured that any relevant minifront changes do not cause
the existing extension to break.