feat(veil): #2527: chart pagination (#2533)
Closes #2515 

Closes #2527 

Added paginated candle request. To make it work properly, we need 2
types of requests:
1. An infinite query that requests pages of 100 candles. When user drags
or zooms out the chart, it loads the next page and appends the data to
the beginning of the candles array. Tradingview charts unfortunately
don't have a method to add partial data to the chart, so I had to play
around with ascending data ordering by candle time.
2. An updatable query for the latest candles, I limited it to the 5
latest. These are refetched on each block.

Also, there's no way to bound the chart vertically by open/close instead
of low/high. I did a trick by limiting the `high` value to be at most 3x
of the `open` candle value. It doesn't go to the moon now, but we can
consider lowering this value even more. Ideally, I think, this needs to
be moved to pindexer as large candles start trust questions about the
chart and generally about Veil


https://github.com/user-attachments/assets/86b1092d-0626-490d-bef4-345eb401f5a9