fix: allow withdrawing withdrawn positions with reserves (#2553)
## Description of Changes

The logic was simply wrong, making the assumption that only closed
positions can be withdrawn.

Instead, this adds a little helper to check if a position is
"fullyWithdrawn", i.e. has no reserves, and is withdrawn.

This is the kind of thing where having an actual domain library would be
helpful.

## Related Issue

Closes #2552.

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.