ci(fly): fast veil deploys by stubbing out wasm  (#2570)
## Description of Changes

I think that should be it to get to fast fly deploys.

This PR:

- deprecates inspect tx in veil, inviting users to use noctis
- create a stub wasm package
- amend the veil-fly dockerfile to surreptitiously swap the wasm package
with wasm-stub

This allows docker to build a complete veil image without needing any
rust or wasm-pack toolchain, and therefore no wasm compilation itself.

I am going to merge this optimistically, if it breaks something we will
revert it.

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.