feat: add git version to veil footer
Adds a git hash in the footer for Veil, so that it's obvious which
commit is running on a given instance. There are multiple frontends
managed by multiple orgs, all on their own deploy schedule, so this
change is intended to make it clear whether a given fix has in fact been
deployed to a given environment.

Closes https://github.com/penumbra-zone/web/issues/2154

lint: prettier