feat: add client side env to ambient context (#2593)
This should fix the social card issues.

I used BASE_URL instead of NEXT_PUBLIC_BASE_URL to avoid the magic env
var, in favor of an explicit one.

## Description of Changes

Please provide a clear and concise description of the changes made in
this PR.
For UI changes, include images or videos to show the visual impact.

## Related Issue

Closes #2575

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.

---------

Co-authored-by: Conor Schaefer <conor@penumbralabs.xyz>