build: permit git dir in docker context
In order to look up commit info during the app build process, the .git
repo dir must be present inside the container. It's debatable whether
this is the shape we want, though: there's a good argument to wrangle
commits outside the build context, and pass that info in via build-args.