build: don't nuke local assets for container build
The "clean" dep was originally introduced to ensure that the container
build context was pure, during debugging. That behavior is solid, thanks
to judicious use of `.dockerignore`, and I routinely run both a dev env
and a container build for veil side by side, and I don't want artifacts
nuked as a side-effect.