feat: add git version to veil footer (#2594)
## Description of Changes

Adds a git hash in the footer for Veil, so that it's obvious which
commit is running on a given instance. There are multiple frontends
managed by multiple orgs, all on their own deploy schedule, so this
change is intended to make it clear whether a given fix has in fact been
deployed to a given environment.

Closes https://github.com/penumbra-zone/web/issues/2154

## Images

It looks like this:


![veil-footer-2](https://github.com/user-attachments/assets/35a759ae-fc09-4d39-9068-76ea065845f7)


## Testing and review

Functionality-wise, this looks "good enough" to me, but I'm mostly
concerned about the shape of the implementation. What you see is mostly
what the LLM thought was a good idea. In particular, please scrutinize
the overriding the node env and the resulting process.env calls; should
we instead use an api env route as described in
https://github.com/penumbra-zone/web/pull/2593?