veil: only display version footer on `/inspect` (#2636)
## Description of Changes

I am not a fan of having a git commit linking to a repository at all,
but let's keep it for now. This isn't something user should see or need
to care about, and the remote URL can be misguiding. It's also not
something that is very important to debate this minute. So for now, this
PR only improves the styling to make it more discreet and moves it to
the nerds page.

## Related Issue

NA

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.