fix: correctly apply subaccount filtering for historical votes in storage (#2650)
## Description of Changes

a conditional check in storage fails when subaccount is 0 (falsy in JS),
causing it to evaluate to false and skip the filtering logic. This in
turns returns the votes for each subaccount, so voting in one account
will incorrectly evaluate each subaccount as already voted. This seems
to have been an issue since the reactor in
https://github.com/penumbra-zone/web/pull/2557 which exposed this issue.
However, this appears to be an edge case that only arises when voting is
initiated from a subaccount first, and then followed by an attempt to
vote with the main account—rather than the other way around.

## Related Issue

https://github.com/penumbra-zone/web/issues/2648

## Checklist Before Requesting Review

- [x] I have ensured that any relevant minifront changes do not cause
the existing extension to break.